feat: abstract class Piece를 상속하는 Knight 클래스 및 테스트 작성
