+++ src/test/java/chess/domain/PieceTest.java
@@ -50,29 +50,6 @@ public class PieceTest {
         assertThat(availableSquares.size()).isEqualTo(14);
     }
 
-    @ParameterizedTest
-    @ValueSource(strings = {"b1", "a2", "b5", "a4", "e4", "d1", "d5", "e2"})
-    @DisplayName("말의 위치(knight)를 받고 말의 종류에 따라 이동할 수 있는 칸 리스트 반환")
-    void calculateScopeKnight(String input) {
-        Piece piece = PieceImpl.of(Color.BLACK, Type.KNIGHT);
-        Set<Square> availableSquares = piece.calculateScope(Square.of("c3"));
-        assertThat(availableSquares.contains(Square.of(input))).isTrue();
-        assertThat(availableSquares.size()).isEqualTo(8);
-    }
-
-    @ParameterizedTest
-    @ValueSource(strings = {"c4", "d5", "d1", "f1", "f5", "g2"})
-    @DisplayName("판의 정보를 가져와서 나이트가 갈 수 있는 칸에 장애물이 있는지 판단하여 이동할 수 있는 리스트 반환하는 테스트")
-    void movableKnightSquareTest(String input) {
-        Map<Square, Piece> board = new HashMap<>();
-        board.put(Square.of("d5"), PieceImpl.of(Color.BLACK, Type.KING));
-        board.put(Square.of("c2"), PieceImpl.of(Color.WHITE, Type.QUEEN));
-        board.put(Square.of("g4"), PieceImpl.of(Color.WHITE, Type.PAWN));
-        Piece piece = PieceImpl.of(Color.WHITE, Type.KNIGHT);
-        Set<Square> availableSquares = piece.calculateMoveBoundary(Square.of("e3"), board);
-        assertThat(availableSquares.contains(Square.of(input))).isTrue();
-        assertThat(availableSquares.size()).isEqualTo(6);
-    }
 
     @ParameterizedTest
     @ValueSource(strings = {"b7", "d7", "b5", "d5", "a4", "e4", "e8"})
